add 64-bit support (closes #249)
list of change details
- fix Xcode project by updating the ARCH build setting. currently ARCH
  is set to ARCHS_STANDARD_32_BIT, by removing ARCH entirely we allow
  Xcode to use its default set of architectures (which currently is
  both 32-bit and 64-bit, but may later change to 64-bit only).
- use casts where NSInteger, NSUInteger and other Cocoa types are mixed
  with primitive types; every cast is explained with a comment
- change variable/parameter types to NSInteger or NSUInteger where it's
  not too much trouble (usually local scope)
- fix format strings to use the "long" modifier (e.g. "%ld" instead of
  "%d"); I did not follow Xcode's suggestion to also cast NSInteger and
  NSUInteger to long and unsigned long because I do not see the need for
  this, and I dislike casts

additional things to consider
- the new 64-bit binaries should have no problem loading the NSCoding
  archive that was created by the 32-bit binaries. at least that's what
  Apple's 64-bit porting guide says.
- it remains to be seen whether the app now requires more memory than
  before.

+++ src/archive/ArchiveViewController.m
@@ -251,7 +251,10 @@ enum DeleteAllSectionItem
     case GamesSection:
     {
       cell = [TableViewCellFactory cellWithType:SubtitleCellType tableView:tableView];
-      ArchiveGame* game = [self.archiveViewModel gameAtIndex:indexPath.row];
+      // Cast is required because NSInteger and int differ in size in 64-bit.
+      // Cast is safe because this app was not made to handle more than
+      // pow(2, 31) files.
+      ArchiveGame* game = [self.archiveViewModel gameAtIndex:(int)indexPath.row];
       cell.textLabel.text = game.name;
       cell.detailTextLabel.text = [@"Last saved: " stringByAppendingString:game.fileDate];
       cell.accessoryType = UITableViewCellAccessoryDisclosureIndicator;
@@ -283,11 +286,14 @@ enum DeleteAllSectionItem
   assert(editingStyle == UITableViewCellEditingStyleDelete);
   if (editingStyle != UITableViewCellEditingStyleDelete)
   {
-    DDLogError(@"%@: Unexpected editingStyle %d", self, editingStyle);
+    DDLogError(@"%@: Unexpected editingStyle %ld", self, editingStyle);
     return;
   }
 
-  ArchiveGame* game = [self.archiveViewModel gameAtIndex:indexPath.row];
+  // Cast is required because NSInteger and int differ in size in 64-bit.
+  // Cast is safe because this app was not made to handle more than
+  // pow(2, 31) files.
+  ArchiveGame* game = [self.archiveViewModel gameAtIndex:(int)indexPath.row];
   DeleteGameCommand* command = [[[DeleteGameCommand alloc] initWithGame:game] autorelease];
   // Temporarily disable KVO observer mechanism so that no table view update
   // is triggered during command execution. Purpose: In a minute, we are going
@@ -313,7 +319,10 @@ enum DeleteAllSectionItem
   {
     case GamesSection:
     {
-      [self viewGame:[self.archiveViewModel gameAtIndex:indexPath.row]];
+      // Cast is required because NSInteger and int differ in size in 64-bit.
+      // Cast is safe because this app was not made to handle more than
+      // pow(2, 31) files.
+      [self viewGame:[self.archiveViewModel gameAtIndex:(int)indexPath.row]];
       break;
     }
     case DeleteAllSection:
