+++ src/command/boardposition/SyncGTPEngineCommand.m
@@ -87,10 +87,10 @@
 - (bool) syncGTPEngineHandicap
 {
   GoGame* game = [GoGame sharedGame];
-  int handicap = game.handicapPoints.count;
+  NSUInteger handicap = game.handicapPoints.count;
   if (0 == handicap)
     return true;
-  GtpCommand* commandFixedHandicap = [GtpCommand command:[NSString stringWithFormat:@"fixed_handicap %d", handicap]];
+  GtpCommand* commandFixedHandicap = [GtpCommand command:[NSString stringWithFormat:@"fixed_handicap %ld", handicap]];
   [commandFixedHandicap submit];
   assert(commandFixedHandicap.response.status);
   return commandFixedHandicap.response.status;
